utils/format/social_media/requirements.txt
requests~=2.32.3
markdownify~=0.12.1
beautifulsoup4~=4.12.3
PyYAML~=6.0.1